lambda/source/requirements.txt
mysql_connector_repackaged==0.3.1
requests==2.26.0